Refactor: Move initialize_database to db/ca.py
I moved the `initialize_database` function from `db.py` to a new file `db/ca.py`.
I adjusted all necessary imports in `db/ca.py`, `db.py`, and `main.py` to reflect this change.
Helper functions used by `initialize_database` (during schema migration) and `seed_initial_data` remain in `db.py` and are imported into `db/ca.py` as needed.
The `main.py` now calls the `initialize_database` function from `db.ca` and continues to use `db.py` for `seed_initial_data` and other DB utility functions.
